Time: 0 ms (100.00%) | Memory: 7.4 MB (73.02%) - LeetSync

+++ 69-sqrtx/sqrtx.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int mySqrt(int x) {
-        long long lo = 0, hi = x;
+        int lo = 0, hi = x;
         int ans = 0;
         while (lo <= hi) {
             long long mid = (lo + hi) / 2;
